HOCS-3949: remove camel consumer + use SqsListener
This change removes the use of camel for the SQS queues. This has a
number of impacts including the removal of a separate queue and DLQ
which should be left to the Terraform config for each environment. This
gives us more control over the usage of the messages we have and allows
us to correctly use the DLQ with effective error handling.