HOCS-3949: rework sns integration
This change reworks the integration with the AWS sdk for SNS. This has
involved organising the properties regarding this into a consistent
manner. This led to a new configuration properties file and model being
added to support the use.

This change also replaces the use of unnecessary beans with these
properties and refactors the camel usage to use the `AmazonSNS` client
directly.