HOCS-3949: rework S3 integration
This change reworks the S3 integration to use spring configuration
property files. This change simplifies the bean creation as if a value
is missing from the config, it will be caught before this. Both
application property files have been changed to accomodate this, with
the main property file having its settings brought under the aws root.