HOCS-3949: camel removal (#41)
This change both improves the resiliency of the service but also 
primarily removes the camel consumer. The overarching change is to move
to using the SQS Listener instead of the Camel queue consumer. This 
removes a layer of abstraction and simplifies the code flow. 

To support this the bean initialisation has been simplified for SQS,
SNS and S3 by removing all unnecessary steps. Integration tests using
LocalStack for SQS, SNS and S3 were also introduced that can both be
run locally or within the pipeline. The retry code has also been 
removed as we prefer a try 1, fail fast approach. There were previously
nuiances with the retry code that was potentially allowing 10 cases to
be created per message. Any dependencies that are no longer required 
have been removed.