Hocs-4376: Upgrade Gradle version to support JDK 17 (#50)
* Hocs-4376: Upgrade Gradle version to support JDK 17

* HOCS-4376: update flaky test

We were mixing JUnit4 and JUnit5 within the test case. This change moved
 back to JUnit4 as the rest of the project is currently using it.

* HOCS-4376: update integration test

There is current known issues when using `@MockBean/@SpyBean` whereby
this bean is not being replaced and the actual bean is being used. This
is causing flaky tests.

This change reworks this test to just ensure that the message is being
read from the queue.

Co-authored-by: Liam MacPherson <liammacpherson1996@gmail.com>
Co-authored-by: Liam MacPherson <11508628+LiamMacP@users.noreply.github.com>