HOCS-4275: remove AuditClient usages
As this service effectively acts as a aggregate service for creating
services, this services responsibility is to send actions to other
services. These other services should handle their respective task and
also the associated auditing. For this reason, auditing within this
service is unnecessary and where applicable replaced with logging.