HOCS-4275: replace/remove auditing with logging (#59)
* HOCS-4275: remove AuditClient usages

As this service effectively acts as a aggregate service for creating
services, this services responsibility is to send actions to other
services. These other services should handle their respective task and
also the associated auditing. For this reason, auditing within this
service is unnecessary and where applicable replaced with logging.

* HOCS-4275: remove sns config for running locally

As the AuditClient was the only use case for the SNS config, this can
now be removed as no longer required.

* HOCS-4275: remove non-local SNS config

As we no longer use the AuditClient, the configuration for the SNS is
no longer required and can be removed.

* HOCS-4275: remove localstack SNS setup

We no longer require SNS setup for Localstack as the service does not
require it. This change removes the SNS setup and changes the setup
script order.

* HOCS-4275: change SQS method name

The method name for the SqsListener was `onAuditEvent`, this is
incorrect as the purpose is to create a complaint [case]. This changes
this method name to the more applicable `onComplaintEvent`.