HOCS-4423: update case data with no correspondent (#61)
When a complaint comes through if there was no correspondent then the
case data was not updated with the complaint type. With all but 1 of the
complaint types correspondents are required. Complaints with
an `EXISTING` type don't have the correspondent requirement and
thus could lead to the case data not having the correct type. This
change ensures that the complaint type is always sent to the casework
service to be updated.