HOCS-4421: use custom clone for branch sha
At the minute we always use the latest version of the code for promoting
to different environments. While this is not an issue in normal cases,
when a change to the deployment config happens this can throw it out of
line. This change ensures that we always check the corresponding branch
out linked to the release, ensuring that we use the config from that
point in time.

This involves disabling the default clone step and adding a custom one
that for pushes uses the global DRONE_COMMIT_SHA environment variable
and VERSION for deployments. By default, if VERSION isn't applicable git
defaults to the head of the main branch on origin. We now promote the
build SHA for this service between environments instead of the build
number/ semvar tag. The semvar tagging infrastructure is still present
to ensure git repositories still have a point of time release.