Update clone step for deployment (#65)
There is issues at the minute whereby the DRONE_COMMIT_SHA variable
isn't being passed down through the events on a push to main. This
change removes the custom clone step in favour of a fetch and checkout.
This change also wraps the environment pass through in quotes to enforce
string coercion.