HOCS-4442: move identities to secret
Presently the userId, teamId and team group are hard-coded within the
repository. While this isn't an issue, this potentially could be used by
a bad actor within the system. This change pulls this out into a
kubernetes secret so that not only is it not in the codebase but is
also easily modifiable if required.