HOCS-4440: send through initial case data
When a case is first created there is a period of time whereby the case
doing have the required data for the Workflow process to process the
case correctly. As we know at the start the complaint type and channel
this can be sent and processed. This allows us to drive the user
experience based on the case is a WebForm.