HOCS-6243: use boolean for message handle disablement
A CSV value was previously used to designate whether a message should be ignored or not. This does not fit the paradigm whereby one deployment should listen to one specific type of message.

This change removes this CSV value in favour of a singular boolean value for message listening disablement.