HOCS-6168: remove multiple handlers code
This change enforces that only a specific handler can be brought in for each listener class, simplifying the logic. This change also includes the addition of an optional MessageType header in the Listener. This should be used to ensure that the right listener is called per message. If the wrong type is specified, an exception is thrown and the respective message is flagged with an appropriate status.