HOCS-6249: add processing code
Add the endpoint, services and related code around the processing of
pending messages that have been added to the message log database.

This first checks the amount of Pending message either between 2 dates
or upto an end date (if not present, the request time is used). If this
count is greater than the maxMessages specified, an
`TooManyMessagesException` is thrown. If the number of messages is under
the threshold, the processing for each message is sequentially
triggered.