HOCS-6249: split user context and message context
As there are now 2 paths through the codebase, there is a need to split
out the current user context and the underlying message context. This
allows for the requests through the API endpoint to have injected
values, while requests outgoing from case-creator (to create a case) can
have the required values.

Correlation IDs are an important part of this methodology, with requests
through the API generating a correlation ID for the lifetime of the
request. For processing of messages outside the API, these generate on a
single message basis.