Change usage of RequestData and remove MessageContext
Previously RequestData was being used for MVC purposes to mint the
correlation ID and store user properties. This has been stripped back to
handling the correlation ID as the processing is only ever instantiated
by a K8 cronjob. Processing during LS can be tracked elsewhere.

As a result of this change of behaviour the MessageContext can also be
removed. All downstream requests use the user and group and this should
be embedded within the clients. The message ID property isn't
appropriate to have within global scope, this is now passed throughout
the application, and as the correlation ID to downstream services to
enable log tracing.